# mydocs/swagger.yaml
openapi: 3.1.1
info:
  title: Learn express and swagger
  description: A course segment about writing docs
  version: 1.0.1
  contact:
    email: natherreran@gmail.com
    url: "https:localhost:4000"

servers:
  - url: "{protocol}://localhost:4000/api/{version}"
    description: localhost version
    variables:
      version:
        enum:
          - v1
          - v2
      protocol: 
        enum:
          - http
          - https
        default: http

components:
  securitySchemes: 
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
    BearerAuth:
      type: http
      scheme: bearer

paths:
  /app:
    get:
      tags:
        - App request - (String) 
      summary: returns a greet message from app
      responses: 
        200:
          description: Successfully
          content: 
            application/json:
              schema: 
                type: string
                example: 'ejemplo de string'
        400:
          description: Bad request
          content: 
            application/json:
              schema: 
                type: string
                example: 'ejemplo de string'
        500:
          description: Internal server error
          content: 
            application/json:
              schema: 
                type: string
                example: 'ejemplo de string'
  /appobject:
    get:
      tags:
        - App request - (Object) 
      summary: returns a unique course
      responses: 
        200:
          description: Successfully
          content: 
            application/json:
              schema: 
                type: object
                properties:
                  id: 
                    type: string
                  name: 
                    type: string
                  price: 
                    type: number
        400:
          description: Bad request
        500:
          description: Internal server error
  /courses:
    get:
      tags:
        - App request - (Array) 
      summary: returns all courses
      responses: 
        200:
          description: Successfully
          content: 
            application/json:
              schema: 
                type: array
                items:
                  type: object
                  properties:
                    id: 
                      type: string
                    name: 
                      type: string
                    price: 
                      type: number
        400:
          description: Bad request
        500:
          description: Internal server error
  /coursequery:
    get:
      tags:
        - App request - (String) 
      summary: returns query
      parameters:
        - name: location
          in: query
          required: true
          schema:
            type: string
            enum: [delhi, london, jaipur]
        - name: device
          in: query
          required: true
          schema:
            type: string
            enum: [web, mobile]
      responses: 
        200:
          description: Successfully
          content: 
            application/json:
              schema: 
                type: object
                properties:
                  location: 
                    type: string
                  device: 
                    type: string
        400:
          description: Bad request
        500:
          description: Internal server error
  /mycourse/{courseID}:
    get:
      tags:
        - App request - (Object) 
      summary: returns course based on request id
      parameters:
        - name: courseID
          in: path
          required: true
          default: 11
          schema:
            type: string
      responses: 
        200:
          description: Successfully
          content: 
            application/json:
              schema: 
                type: object
                properties:
                  id: 
                    type: string
                  name: 
                    type: string
                  price: 
                    type: number
        400:
          description: Bad request
        500:
          description: Internal server error
  /addcourse:
    post:
      tags:
        - App request - (Object) 
      summary: add a new course to existing courses
      consumes:
        - application/json
      produces:
        - application/json
      requestBody:
        required: true
        content:
          application/json:
            schema: 
              type: object
              properties:
                id: 
                  type: string
                name: 
                  type: string
                price: 
                  type: number
      responses: 
        200:
          description: Successfully
          content: 
            application/json:
              schema: 
                type: boolean
        400:
          description: Bad request
        500:
          description: Internal server error     
  /courseupload:
    post:
      tags:
        - App request - (String) 
      summary: uploading course image
      consumes:
        - application/json
      produces:
        - application/json
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: 
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses: 
        200:
          description: Successfully
        400:
          description: Bad request
        500:
          description: Internal server error
      